requirements.txt
piexif
Pillow